requirements.txt
flask
opencv-python
easyocr
pillow
numpy
openai
